msm: pil-dsps: Combine SSR and PIL drivers
Combine both drivers into one so that we can merge the PIL and
SSR frameworks. The ramdump feature is not supported on 8660 so
just copy the 8960 version of the segments for now. In the future
we'll add an API to query the segments from PIL so that changes
in the image between platforms can be handled properly.

Change-Id: I239939457fb2fc8023a5923220f52dc6e099aa85
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>