msm: pil-q6v4-mss: Combine SSR and PIL drivers
Combine both drivers into one so that we can merge the PIL and
SSR frameworks. To cope with the differences between 8960 and
9615 (the former supports PIL while the latter does not) we check
to see if there is any platform data specified for the device. If
there's platform data we know we're driving a PIL supported
device and act accordingly. On 9615, we only support calling
panic on modem watchdog interrupts, so it's safe to avoid
ioremapping registers.

Change-Id: I6a530fb640b295448805bfeb9d5beee389bcf6f2
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>