treewide: Replace pil_get()/put() with subsystem_get()/put()
Do a tree wide sed to replace pil_get and pil_put() with their
subsystem counterpart. With all the pil drivers migrated over to
the new subsystem start/stop API this is essentially a no-op.

We rename "q6" to "adsp" to be more in line with how future
chipsets name the low power audio subsystem. This requires minor
changes to userspace; hopefully no more changes will be
necessary.

Change-Id: I22dbbdcf7f5aaaa80394a6d70b5cfc24cdb32044
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>