platform: msm: Add software workaround to cover PM8941 LPG bug
In PM8941 PMIC chip, there is a hardware bug in the PWM module
design so that if a PWM write results in input of logic 0 followed
by a pwm_en = 0 then the subsequent write of PWM value will not be
updated to the output.

Add an additional PWM write before enabling PWM so that PWM value
is updated at the output.

Change-Id: I8e7f6c6892b6ffd3ee78fa4d1e6899f77bdef789
Signed-off-by: Jay Chokshi <jchokshi@codeaurora.org>