qup_i2c: keep the ahb clock on for SGLTE 8960 only
Don't keep the AHB clock on for all types of 8960 targets
as the GSBI4 is shared by I2C and UART cores from different
processors on only SGLTE 8960 type platform. If the ahb clock
is kept on for GSBI4 from apps side, this needs a change in UART
driver on modem side to avoid touching the ahb clock and may
lead to some i2c errors if the modem doesn't have this change.
Hence we are keeping the ahb clock on for only SGLTE 8960 target
type.

CRs-fixed: 387908
Change-Id: If032d48a2cc0c9a9658359f05fe054d2e1f758c4
Signed-off-by: Kiran Gunda <kgunda@codeaurora.org>