arm/dt: 8974: Update SPM sequences for L2 retention
Remove the L2 SPM command that waits for cpus to go into low power mode
in the retention sequence. The command prevents a surge in current into
L2 when the cores are being power collapsed prior to turning off the L2
logic. The L2 logic is not turned off during retention and hence this
command isn't needed for retention sequence.

Change-Id: I1822bba99d6f7010092bc77ed2eddd0dccffd7be
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>