hwmon: epm_adc: Add devicetree binding document
Embedded power measurement (EPM) uses a PSoC5
microcontroller by Cyprus to issue ADC request
to read voltage and current measurements.

The device node adds EPM as part of the SPI
devicetree node since the communication between
the MSM and PSoC is over SPI. The required properties
for the EPM node are the number of channels supported,
gain, rsense and the channel type to distinguish
between voltage and current channels.

Change-Id: Ifdd554c0d6fb36d96b20582f449d48c20f3bcee4
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>